Testing: making add user file
